requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
deep_translator==1.11.4
pdfplumber==0.10.3
protobuf==4.25.2
PyJWT==2.8.0
pymantic==1.0.0
PyPDF2==3.0.1
python_docx==1.1.0
reportlab==4.0.9
Requests==2.31.0
shortuuid==1.0.11
spacy==3.7.2
wget==3.2
Django==4.2.7
rdflib==6.0.2
pymantic==1.0.0
websockets==12.0.0
# django-channels==0.7.0 (required only for docker )
channels==3.0.5
django-cors-headers==4.3.1
django-allauth==0.59.0
spacy==3.7.2
deep-translator==1.11.4
google-auth==2.25.2
mysqlclient==2.2.1
pymysql==1.1.0
tzlocal==5.2
schedule==1.2.1
